// src/infrastructure/interfaces/movies-response.ts
// intermedio entre respuesta de moviedb y modelo

export interface MoviesResponse {
    dates: Dates;
    page: number;
    results: Result[];
    total_pages: number;
    total_results: number;
}

export interface Dates {
    maximum: Date;
    minimum: Date;
}

export interface Result {
    adult: boolean;
    backdrop_path: string;
    genre_ids: number[];
    id: number;
    title: string;
    original_language: string;
    original_title: string;
    overview: string;
    popularity: number;
    poster_path: string;
    release_date: Date;
    softcore: boolean;
    video: boolean;
    vote_average: number;
    vote_count: number;
}